Hopper is in
Made a container called hopper which is passed into
every new class that would use static access.
Simple rendering manager that has items which draw.
component::Render creates its own items and sets
a lambda to be the draw function.
This lambda might will capture the render component
so that it will be able to get the owning entities
possition, etc