Temp commit
Lots of work has been done...
but i have found that my approach is wrong.
The pointers to items in vectors wont be correct if the vector
shuffles the memory around (which can be done when the list
grows).
New idea, return the make_shared ptr, then just grab the raw
pointer from it and put that in the list.
Then have the deleter for the shared_ptr remove it from the list
before it gets deleted